[vector] Rename vector index module
