+++ pom.xml
@@ -76,7 +76,7 @@ under the License.
         <module>paimon-vortex</module>
         <module>paimon-mosaic</module>
         <module>paimon-tantivy</module>
-        <module>paimon-ivfpq</module>
+        <module>paimon-vector</module>
     </modules>
 
     <properties>
